1. bash

+++ jvbo-springboot/jvbo-springboot-practice/doc/simple-start-dev.sh
@@ -1,6 +1,10 @@
 #/bin/sh
 
+if [ $(lsof -t -i:8001) ]; then
 kill -9 $(lsof -t -i:8001)
+else
+echo "no 8001"
+fi
 
 mv jvbo-springboot-practice-0.0.1-SNAPSHOT.war jvbo-springboot-practice-0.0.1-SNAPSHOT.war`date '+%Y-%m-%d-%H%M%S'`
